apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
webauthn==2.0.0
httpx==0.26.0
scikit-learn==1.4.0
numpy==1.26.3
aiofiles==23.2.1
